Moved 'cached_block_allocator' to detail
This now privatizes the cached_block_allocator for internal use
only.

Before this change, it was always possible to wrap amy block allocator
in a cached one; including other cached block allocators. This internalization
leaves the responsibility up to the block-allocator author to support
caching.
