+++ include/bit/memory/block_allocators/detail/cached_block_allocator.inl
@@ -6,10 +6,10 @@
 //----------------------------------------------------------------------------
 
 template<typename BlockAllocator>
-template<typename...Args>
-inline bit::memory::cached_block_allocator<BlockAllocator>
-  ::cached_block_allocator( Args&&...args )
-  : m_storage( std::forward<Args>(args)... )
+template<typename Arg0, typename...Args, typename>
+inline bit::memory::detail::cached_block_allocator<BlockAllocator>
+  ::cached_block_allocator( Arg0&& arg0, Args&&...args )
+  : BlockAllocator( std::forward<Arg0>(arg0), std::forward<Args>(args)... )
 {
 
 }
@@ -17,12 +17,12 @@ inline bit::memory::cached_block_allocator<BlockAllocator>
 //----------------------------------------------------------------------------
 
 template<typename BlockAllocator>
-inline bit::memory::cached_block_allocator<BlockAllocator>
+inline bit::memory::detail::cached_block_allocator<BlockAllocator>
   ::~cached_block_allocator()
 {
   // Deallocate every allocated block
-  while( !m_storage.cache.empty() ) {
-    m_storage.deallocate_block( m_storage.cache.request_block() );
+  while( !m_cache.empty() ) {
+    BlockAllocator::deallocate_block( m_cache.request_block() );
   }
 }
 
@@ -32,18 +32,18 @@ inline bit::memory::cached_block_allocator<BlockAllocator>
 
 template<typename BlockAllocator>
 inline bit::memory::owner<bit::memory::memory_block>
-  bit::memory::cached_block_allocator<BlockAllocator>::allocate_block()
+  bit::memory::detail::cached_block_allocator<BlockAllocator>::allocate_block()
 {
-  return m_storage.cache.request_block( m_storage );
+  return m_cache.request_block( static_cast<BlockAllocator&>(*this) );
 }
 
 //----------------------------------------------------------------------------
 
 template<typename BlockAllocator>
-inline void bit::memory::cached_block_allocator<BlockAllocator>
+inline void bit::memory::detail::cached_block_allocator<BlockAllocator>
   ::deallocate_block( owner<memory_block> block )
 {
-  m_storage.cache.store_block( std::move(block) );
+  m_cache.store_block( std::move(block) );
 }
 
 #endif /* BIT_MEMORY_BLOCK_ALLOCATORS_DETAIL_CACHED_BLOCK_ALLOCATOR_INL */
