+++ include/bit/memory/block_allocators/thread_local_block_allocator.hpp
@@ -58,7 +58,6 @@ namespace bit {
       using block_size      = std::integral_constant<std::size_t,BlockSize>;
       using block_alignment = std::integral_constant<std::size_t,Align>;
 
-
       //----------------------------------------------------------------------
       // Constructor
       //----------------------------------------------------------------------
